pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "image-vector-search"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1",
  "itsdangerous>=2.2,<3",
  "fastmcp>=3,<4",
  "httpx>=0.28,<1",
  "milvus-lite>=2.4,<3",
  "pillow>=11,<12",
  "pydantic-settings>=2.8,<3",
  "pymilvus>=2.5,<3",
  "python-multipart>=0.0.9,<1",
  "setuptools>=75,<76",
  "uvicorn[standard]>=0.34,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.26,<1",
  "respx>=0.22,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
image_vector_search = [
  "frontend/dist/**",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning:milvus_lite.*:",
]
